+++ spec/rubocop/cop/style/string_literals_spec.rb
@@ -99,6 +99,11 @@ describe RuboCop::Cop::Style::StringLiterals, :config do
       expect(cop.offenses).to be_empty
     end
 
+    it 'accepts double quotes when unicode control sequence is used' do
+      inspect_source(cop, '"Espa\u00f1a"')
+      expect(cop.offenses).to be_empty
+    end
+
     it 'accepts double quotes at the start of regexp literals' do
       inspect_source(cop, 's = /"((?:[^\\"]|\\.)*)"/')
       expect(cop.offenses).to be_empty
@@ -162,14 +167,14 @@ describe RuboCop::Cop::Style::StringLiterals, :config do
                                   'special symbols.'])
     end
 
-    it 'registers an offense for hello... using hex escapes' do
-      inspect_source(cop, '"\\x68\\x65\\x6c\\x6c\\x6f"')
+    it 'registers an offense for words with non-ascii chars' do
+      inspect_source(cop, '"España"')
       expect(cop.offenses.size).to eq(1)
     end
 
-    it 'autocorrects hello... using hex escapes' do
-      new_source = autocorrect_source(cop, '"\\x68\\x65\\x6c\\x6c\\x6f"')
-      expect(new_source).to eq("'hello'")
+    it 'autocorrects words with non-ascii chars' do
+      new_source = autocorrect_source(cop, '"España"')
+      expect(new_source).to eq("'España'")
     end
   end